move: 버전별 디렉토리 삭제, 한 곳에 모아두도록 옮김
