+++ product-service/src/main/java/com/monkey/productservice/application/service/ProductServiceApiV4.java
@@ -1,4 +1,4 @@
-package com.monkey.productservice.application.service.v4;
+package com.monkey.productservice.application.service;
 
 import com.monkey.common_module.aop.AccessLevel;
 import com.monkey.common_module.aop.CheckUserRole;
